+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/components/business/common-controlling-members.component.ts
@@ -1,5 +1,16 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { FormArray, FormBuilder, FormControl, FormGroup } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
+import { MatTableDataSource } from '@angular/material/table';
+import { LicenceResponse } from '@app/api/models';
+import { showHideTriggerSlideAnimation } from '@app/core/animations';
+import { BooleanTypeCode } from '@app/core/code-types/model-desc.models';
+import { SPD_CONSTANTS } from '@app/core/constants/constants';
+import { DialogComponent, DialogOptions } from '@app/shared/components/dialog.component';
 import { BusinessApplicationService } from '../../services/business-application.service';
+import { LicenceChildStepperStepComponent } from '../../services/licence-application.helper';
+import { LookupSwlDialogData, ModalLookupSwlComponent } from './modal-lookup-swl.component';
+import { ModalMemberWithoutSwlEditComponent } from './modal-member-without-swl-edit.component';
 
 @Component({
 	selector: 'app-common-controlling-members',
@@ -30,28 +41,418 @@ import { BusinessApplicationService } from '../../services/business-application.
 					<mat-panel-title>Controlling Members</mat-panel-title>
 				</mat-expansion-panel-header>
 
-				<app-common-business-controlling-members
-					[form]="controllingMembersFormGroup"
-				></app-common-business-controlling-members>
+				<form [formGroup]="form" novalidate>
+					<div class="row mt-4" *ngIf="controllingMembersWithSwlExist">
+						<div class="col-12">
+							<div class="mb-2 text-primary-color">Controlling Members with a Security Worker Licence:</div>
+							<mat-table [dataSource]="dataSourceWithSWL">
+								<ng-container matColumnDef="licenceHolderName">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>Full Name</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Full Name:</span>
+										{{ member.licenceHolderName | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="licenceNumber">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>
+										Security Worker Licence Number
+									</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Security Worker Licence Number:</span>
+										{{ member.licenceNumber | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="licenceStatusCode">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>Licence Status</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Status:</span>
+										{{ member.licenceStatusCode | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="expiryDate">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>Expiry Date</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Expiry Date:</span>
+										{{ member.expiryDate | formatDate | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="action1">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef></mat-header-cell>
+									<mat-cell *matCellDef="let member; let i = index">
+										<button
+											mat-flat-button
+											class="table-button w-auto"
+											style="color: var(--color-red);"
+											aria-label="Remove controlling member"
+											(click)="onRemoveMember(true, i)"
+										>
+											<mat-icon>delete_outline</mat-icon>Remove
+										</button>
+									</mat-cell>
+								</ng-container>
+
+								<mat-header-row *matHeaderRowDef="columnsWithSWL; sticky: true"></mat-header-row>
+								<mat-row class="mat-data-row" *matRowDef="let row; columns: columnsWithSWL"></mat-row>
+							</mat-table>
+						</div>
+					</div>
+
+					<div class="row mt-3" *ngIf="!isMaxNumberOfControllingMembers">
+						<ng-container *ngIf="!controllingMembersWithSwlExist">
+							<div class="mb-2">No controlling members with a Security Worker Licence exist</div>
+						</ng-container>
+						<div class="col-md-12" [ngClass]="isWizard ? 'col-lg-7 col-xl-6' : 'col-lg-6 col-xl-5'">
+							<button mat-flat-button color="primary" class="large mb-2" (click)="onAddMemberWithSWL()">
+								Add Member with a Security Worker Licence
+							</button>
+						</div>
+					</div>
+
+					<mat-divider class="mat-divider-primary my-2"></mat-divider>
+
+					<div class="row mt-4" *ngIf="controllingMembersWithoutSwlExist">
+						<div class="col-12">
+							<div class="text-primary-color">Controlling Members without a Security Worker Licence:</div>
+							<mat-table [dataSource]="dataSourceWithoutSWL">
+								<ng-container matColumnDef="licenceHolderName">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>Full Name</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Full Name:</span>
+										{{ member.licenceHolderName | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="clearanceStatus">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef>
+										Controlling Member Clearance
+									</mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<span class="mobile-label">Controlling Member Clearance:</span>
+										{{ member.clearanceStatus | default }}
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="action1">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef></mat-header-cell>
+									<mat-cell *matCellDef="let member">
+										<button
+											mat-flat-button
+											class="table-button"
+											style="color: var(--color-green);"
+											aria-label="Edit controlling member"
+											(click)="onEditMemberWithoutSWL(member)"
+											*ngIf="!member.licenceNumber"
+										>
+											<mat-icon>edit</mat-icon>Edit
+										</button>
+									</mat-cell>
+								</ng-container>
+
+								<ng-container matColumnDef="action2">
+									<mat-header-cell class="mat-table-header-cell" *matHeaderCellDef></mat-header-cell>
+									<mat-cell *matCellDef="let member; let i = index">
+										<button
+											mat-flat-button
+											class="table-button w-auto"
+											style="color: var(--color-red);"
+											aria-label="Remove controlling member"
+											(click)="onRemoveMember(false, i)"
+										>
+											<mat-icon>delete_outline</mat-icon>Remove
+										</button>
+									</mat-cell>
+								</ng-container>
+
+								<mat-header-row *matHeaderRowDef="columnsWithoutSWL; sticky: true"></mat-header-row>
+								<mat-row class="mat-data-row" *matRowDef="let row; columns: columnsWithoutSWL"></mat-row>
+							</mat-table>
+						</div>
+					</div>
+
+					<div class="row mt-3">
+						<ng-container *ngIf="!controllingMembersWithoutSwlExist">
+							<div class="mb-2">No controlling members without a Security Worker Licence exist</div>
+						</ng-container>
+						<ng-container *ngIf="isMaxNumberOfControllingMembers; else CanAddMember2">
+							<div class="col-12">
+								<app-alert type="warning" icon="warning">
+									<div>The maximum number of controlling members has been reached</div>
+								</app-alert>
+							</div>
+						</ng-container>
+						<ng-template #CanAddMember2>
+							<div class="col-md-12" [ngClass]="isWizard ? 'col-lg-7 col-xl-6' : 'col-lg-6 col-xl-5'">
+								<button mat-flat-button color="primary" class="large mb-2" (click)="onAddMemberWithoutSWL()">
+									Add Member without a Security Worker Licence
+								</button>
+							</div>
+						</ng-template>
+					</div>
+
+					<div *ngIf="allowDocumentUpload" @showHideTriggerSlideAnimation>
+						<div class="row mt-2">
+							<div class="col-12">
+								<mat-divider class="mat-divider-main my-3"></mat-divider>
+								<div class="text-minor-heading lh-base mb-2">
+									Upload a copy of the corporate registry documents for your business in the province in which you are
+									originally registered <span class="optional-label">(optional)</span>
+								</div>
+								<app-file-upload
+									(fileUploaded)="onFileUploaded($event)"
+									(fileRemoved)="onFileRemoved()"
+									[control]="attachments"
+									[maxNumberOfFiles]="10"
+									[files]="attachments.value"
+								></app-file-upload>
+								<mat-error
+									class="mat-option-error"
+									*ngIf="
+										(form.get('attachments')?.dirty || form.get('attachments')?.touched) &&
+										form.get('attachments')?.invalid &&
+										form.get('attachments')?.hasError('required')
+									"
+									>This is required</mat-error
+								>
+							</div>
+						</div>
+					</div>
+				</form>
 			</mat-expansion-panel>
 		</mat-accordion>
 	`,
-	styles: [],
+	styles: [
+		`
+			.mat-column-action1 {
+				min-width: 150px;
+				max-width: 150px;
+				.table-button {
+					min-width: 130px;
+				}
+			}
+			.mat-column-action2 {
+				min-width: 150px;
+				max-width: 150px;
+				.table-button {
+					min-width: 130px;
+				}
+			}
+		`,
+	],
+	animations: [showHideTriggerSlideAnimation],
 })
-export class CommonControllingMembersComponent {
-	controllingMembersFormGroup = this.businessApplicationService.controllingMembersFormGroup;
+export class CommonControllingMembersComponent implements OnInit, LicenceChildStepperStepComponent {
+	booleanTypeCodes = BooleanTypeCode;
+	maxNumberOfMembers = SPD_CONSTANTS.maxCount.controllingMembers;
+
+	form = this.businessApplicationService.controllingMembersFormGroup;
 
 	@Input() defaultExpanded = false;
+	@Input() isWizard = false;
 
-	constructor(private businessApplicationService: BusinessApplicationService) {}
+	isBcBusinessAddress = true;
+	allowDocumentUpload = false;
 
-	// get isNewOrRenewal(): boolean {
-	// 	return (
-	// 		this.applicationTypeCode === ApplicationTypeCode.Renewal || this.applicationTypeCode === ApplicationTypeCode.New
-	// 	);
-	// }
+	dataSourceWithSWL!: MatTableDataSource<any>;
+	columnsWithSWL: string[] = ['licenceHolderName', 'licenceNumber', 'licenceStatusCode', 'expiryDate', 'action1'];
 
-	// get isUpdate(): boolean {
-	// 	return this.applicationTypeCode === ApplicationTypeCode.Update;
-	// }
+	dataSourceWithoutSWL!: MatTableDataSource<any>;
+	columnsWithoutSWL: string[] = ['licenceHolderName', 'clearanceStatus', 'action1', 'action2'];
+
+	constructor(
+		private formBuilder: FormBuilder,
+		private dialog: MatDialog,
+		private businessApplicationService: BusinessApplicationService
+	) {}
+
+	ngOnInit(): void {
+		this.isBcBusinessAddress = this.businessApplicationService.isBcBusinessAddress();
+
+		this.dataSourceWithSWL = new MatTableDataSource(this.membersWithSwlList.value);
+		this.dataSourceWithoutSWL = new MatTableDataSource(this.membersWithoutSwlList.value);
+	}
+
+	isFormValid(): boolean {
+		return true;
+	}
+
+	onRemoveMember(isWithSwl: boolean, index: number) {
+		const data: DialogOptions = {
+			icon: 'warning',
+			title: 'Confirmation',
+			message: 'Are you sure you want to remove this member?',
+			actionText: 'Yes, remove',
+			cancelText: 'Cancel',
+		};
+
+		this.dialog
+			.open(DialogComponent, { data })
+			.afterClosed()
+			.subscribe((response: boolean) => {
+				this.controllingMemberChanged();
+
+				if (response) {
+					if (isWithSwl) {
+						this.membersWithSwlList.removeAt(index);
+						this.dataSourceWithSWL = new MatTableDataSource(this.membersWithSwlList.value);
+					} else {
+						this.membersWithoutSwlList.removeAt(index);
+						this.dataSourceWithoutSWL = new MatTableDataSource(this.membersWithoutSwlList.value);
+					}
+				}
+			});
+	}
+
+	onAddMemberWithSWL(): void {
+		const dialogOptions: LookupSwlDialogData = {
+			title: 'Add Member with Security Worker Licence',
+			notValidSwlMessage: `'Cancel' to exit this dialog and then add them as a member without a security worker licence to proceed.`,
+		};
+		this.dialog
+			.open(ModalLookupSwlComponent, {
+				width: '800px',
+				data: dialogOptions,
+			})
+			.afterClosed()
+			.subscribe((resp: any) => {
+				const memberData: LicenceResponse = resp?.data;
+				if (memberData) {
+					this.controllingMemberChanged();
+
+					this.membersWithSwlList.push(this.newMemberRow(memberData));
+					this.dataSourceWithSWL.data = this.membersWithSwlList.value;
+				}
+			});
+	}
+
+	onEditMemberWithoutSWL(member: any): void {
+		this.memberDialogWithoutSWL(member, false);
+	}
+
+	onAddMemberWithoutSWL(): void {
+		this.memberDialogWithoutSWL({}, true);
+	}
+
+	onFileUploaded(_file: File): void {
+		// TODO upload file on partial save
+		this.businessApplicationService.hasValueChanged = true;
+
+		if (this.businessApplicationService.isAutoSave()) {
+			// this.businessApplicationService.addUploadDocument(LicenceDocumentTypeCode.xxx, file).subscribe({
+			// 	next: (resp: any) => {
+			// 		const matchingFile = this.attachments.value.find((item: File) => item.name == file.name);
+			// 		matchingFile.documentUrlId = resp.body[0].documentUrlId;
+			// 	},
+			// 	error: (error: any) => {
+			// 		console.log('An error occurred during file upload', error);
+			// 		this.hotToastService.error('An error occurred during the file upload. Please try again.');
+			// 		this.fileUploadComponent.removeFailedFile(file);
+			// 	},
+			// });
+		}
+	}
+
+	onFileRemoved(): void {
+		this.businessApplicationService.hasValueChanged = true;
+	}
+
+	private controllingMemberChanged(): void {
+		if (this.isBcBusinessAddress) return;
+
+		this.allowDocumentUpload = true;
+	}
+
+	private memberDialogWithoutSWL(dialogOptions: any, isCreate: boolean): void {
+		this.dialog
+			.open(ModalMemberWithoutSwlEditComponent, {
+				width: '800px',
+				data: dialogOptions,
+			})
+			.afterClosed()
+			.subscribe((resp: any) => {
+				const memberData = resp?.data;
+				if (memberData) {
+					if (isCreate) {
+						this.controllingMemberChanged();
+						this.membersWithoutSwlList.push(this.newMemberRow(memberData));
+					} else {
+						const memberIndex = this.membersWithoutSwlList.value.findIndex(
+							(item: any) => item.bizContactId == dialogOptions.bizContactId!
+						);
+						this.patchMemberData(memberIndex, memberData);
+					}
+
+					this.dataSourceWithoutSWL.data = this.membersWithoutSwlList.value;
+				}
+			});
+	}
+
+	private newMemberRow(memberData: any): FormGroup {
+		return this.formBuilder.group({
+			licenceHolderName: [memberData.licenceHolderName ?? `${memberData.givenName} ${memberData.surname}`],
+			bizContactId: null,
+			contactId: [memberData.licenceHolderId],
+			givenName: [memberData.givenName],
+			middleName1: [memberData.middleName1],
+			middleName2: [memberData.middleName2],
+			surname: [memberData.surname],
+			emailAddress: [memberData.emailAddress],
+			noEmailAddress: [memberData.noEmailAddress],
+			phoneNumber: [memberData.phoneNumber],
+			licenceId: [memberData.licenceId],
+			licenceNumber: [memberData.licenceNumber],
+			licenceStatusCode: [memberData.licenceStatusCode],
+			licenceTermCode: [memberData.licenceTermCode],
+			expiryDate: [memberData.expiryDate],
+			clearanceStatus: [memberData.clearanceStatus],
+		});
+	}
+
+	private patchMemberData(memberIndex: number, memberData: any) {
+		if (memberIndex < 0) {
+			return;
+		}
+
+		this.membersWithoutSwlList.at(memberIndex).patchValue({
+			licenceHolderName: memberData.licenceHolderName ?? `${memberData.givenName} ${memberData.surname}`,
+			bizContactId: memberData.bizContactId,
+			givenName: memberData.givenName,
+			middleName1: memberData.middleName1,
+			middleName2: memberData.middleName2,
+			surname: memberData.surname,
+			emailAddress: memberData.emailAddress,
+			noEmailAddress: memberData.noEmailAddress,
+			licenceId: memberData.licenceId,
+			licenceNumber: memberData.licenceNumber,
+			licenceStatusCode: memberData.licenceStatusCode,
+			licenceTermCode: memberData.licenceTermCode,
+			expiryDate: memberData.expiryDate,
+			clearanceStatus: memberData.clearanceStatus,
+		});
+	}
+
+	get attachments(): FormControl {
+		return this.form.get('attachments') as FormControl;
+	}
+	get membersWithSwlList(): FormArray {
+		return <FormArray>this.form.get('membersWithSwl');
+	}
+	get membersWithoutSwlList(): FormArray {
+		return <FormArray>this.form.get('membersWithoutSwl');
+	}
+	get controllingMembersWithSwlExist(): boolean {
+		return this.dataSourceWithSWL.data.length > 0;
+	}
+	get controllingMembersWithoutSwlExist(): boolean {
+		return this.dataSourceWithoutSWL.data.length > 0;
+	}
+	get numberOfControllingMembers(): number {
+		return this.dataSourceWithSWL.data.length + this.dataSourceWithoutSWL.data.length;
+	}
+	get isMaxNumberOfControllingMembers(): boolean {
+		return this.numberOfControllingMembers >= SPD_CONSTANTS.maxCount.controllingMembers;
+	}
 }
