+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/components/business/modal-member-without-swl-edit.component.ts
@@ -68,7 +68,10 @@ import { FormErrorStateMatcher } from '@app/shared/directives/form-error-state-m
 					<button mat-stroked-button mat-dialog-close class="large" color="primary">Cancel</button>
 				</div>
 				<div class="offset-md-4 col-md-4 col-sm-12 mb-2">
-					<button mat-flat-button color="primary" class="large" (click)="onSave()">Apply</button>
+					<button mat-flat-button color="primary" class="large" (click)="onSave()">
+						<span *ngIf="!isEdit">Add</span>
+						<span *ngIf="isEdit">Update</span>
+					</button>
 				</div>
 			</div>
 		</mat-dialog-actions>
@@ -94,8 +97,8 @@ export class ModalMemberWithoutSwlEditComponent implements OnInit {
 	ngOnInit(): void {
 		this.form.reset();
 		this.form.patchValue(this.dialogData);
-		this.isEdit = !!this.dialogData.id;
-		this.title = this.dialogData.id
+		this.isEdit = !!this.dialogData.bizContactId;
+		this.title = this.isEdit
 			? 'Edit Member without Security Worker Licence'
 			: 'Add Member without Security Worker Licence';
 	}
