+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/licence-application.module.ts
@@ -92,8 +92,6 @@ import { BusinessManagersComponent } from './components/business/business-manage
 import { BusinessProfileComponent } from './components/business/business-profile.component';
 import { BusinessUserApplicationsComponent } from './components/business/business-user-applications.component';
 import { CommonBusinessBcBranchesComponent } from './components/business/common-business-bc-branches.component';
-import { CommonBusinessControllingMembersComponent } from './components/business/common-business-controlling-members.component';
-import { CommonBusinessEmployeesComponent } from './components/business/common-business-employees.component';
 import { CommonBusinessInformationComponent } from './components/business/common-business-information.component';
 import { CommonBusinessProfileComponent } from './components/business/common-business-profile.component';
 import { CommonControllingMembersComponent } from './components/business/common-controlling-members.component';
@@ -227,8 +225,6 @@ import { LicenceApplicationService } from './services/licence-application.servic
 		CommonBcDriverLicenceComponent,
 		CommonBusinessAddressComponent,
 		CommonBusinessBcBranchesComponent,
-		CommonBusinessControllingMembersComponent,
-		CommonBusinessEmployeesComponent,
 		CommonBusinessInformationComponent,
 		CommonBusinessProfileComponent,
 		CommonBusinessTermsComponent,
